+++ ruby/ruby/Ice.rb
@@ -237,7 +237,7 @@ require_relative 'IceLocal/Current.rb'
 require_relative 'IceLocal/Communicator.rb'
 require_relative 'IceLocal/EndpointSelectionType.rb'
 require 'Ice/EndpointTypes.rb'
-require_relative 'IceLocal/LocalException.rb'
+require_relative 'IceLocal/LocalExceptions.rb'
 require 'Ice/Locator.rb'
 require_relative 'IceLocal/Logger.rb'
 require_relative 'IceLocal/ValueFactory.rb'
