+++ ruby/ruby/IceLocal/LocalExceptions.rb
